requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
betfairlightweight>=2.23.0
requests>=2.28.0
openpyxl>=3.1.0
xlrd>=2.0.1
matplotlib>=3.7.0
scipy>=1.10.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.65.0
shap>=0.44.0
